Modify loading and error screen layout
